move local_computation decorator from protocol to main namespace
